Refactoring and restructuring; small changes.
